requirements.txt
python-telegram-bot==22.0
python-dotenv==1.1.0
nest-asyncio==1.6.0
